Update audit log event data (#61769)
