+++ package.json
@@ -97,6 +97,7 @@
     "start-for-playwright": "cross-env ROOT=src/fixtures/fixtures TRANSLATIONS_FIXTURE_ROOT=src/fixtures/fixtures/translations ENABLED_LANGUAGES=en,ja NODE_ENV=test tsx src/frame/server.ts",
     "symlink-from-local-repo": "tsx src/early-access/scripts/symlink-from-local-repo.ts",
     "sync-audit-log": "tsx src/audit-logs/scripts/sync.ts",
+    "rebuild-audit-log-dedup": "tsx src/audit-logs/scripts/rebuild-dedup.ts",
     "sync-codeql-cli": "tsx src/codeql-cli/scripts/sync.ts",
     "sync-graphql": "tsx src/graphql/scripts/sync.ts",
     "sync-rest": "tsx src/rest/scripts/update-files.ts",
