+++ src/audit-logs/data/ghec/enterprise.json
@@ -23441,6 +23441,7 @@
       "query_suite",
       "threat_model",
       "languages",
+      "languages_removed",
       "public_repo",
       "created_at",
       "operation_type",
