+++ src/audit-logs/data/ghec/organization.json
@@ -18023,6 +18023,7 @@
       "query_suite",
       "threat_model",
       "languages",
+      "languages_removed",
       "public_repo",
       "created_at",
       "operation_type",
