+++ src/audit-logs/data/ghes-3.17/enterprise.json
@@ -17456,6 +17456,7 @@
       "query_suite",
       "threat_model",
       "languages",
+      "languages_removed",
       "public_repo",
       "created_at",
       "operation_type",
