+++ src/audit-logs/data/ghes-3.17/organization.json
@@ -16781,6 +16781,7 @@
       "query_suite",
       "threat_model",
       "languages",
+      "languages_removed",
       "public_repo",
       "created_at",
       "operation_type",
