+++ src/audit-logs/data/ghes-3.18/enterprise.json
@@ -17526,6 +17526,7 @@
       "query_suite",
       "threat_model",
       "languages",
+      "languages_removed",
       "public_repo",
       "created_at",
       "operation_type",
