+++ src/audit-logs/data/ghes-3.18/organization.json
@@ -17061,6 +17061,7 @@
       "query_suite",
       "threat_model",
       "languages",
+      "languages_removed",
       "public_repo",
       "created_at",
       "operation_type",
