+++ src/audit-logs/data/ghes-3.19/enterprise.json
@@ -18483,6 +18483,7 @@
       "query_suite",
       "threat_model",
       "languages",
+      "languages_removed",
       "public_repo",
       "created_at",
       "operation_type",
