+++ src/audit-logs/data/ghes-3.19/organization.json
@@ -17774,6 +17774,7 @@
       "query_suite",
       "threat_model",
       "languages",
+      "languages_removed",
       "public_repo",
       "created_at",
       "operation_type",
