+++ src/audit-logs/data/ghes-3.20/enterprise.json
@@ -18516,6 +18516,7 @@
       "query_suite",
       "threat_model",
       "languages",
+      "languages_removed",
       "public_repo",
       "created_at",
       "operation_type",
