+++ src/audit-logs/data/ghes-3.20/organization.json
@@ -17711,6 +17711,7 @@
       "query_suite",
       "threat_model",
       "languages",
+      "languages_removed",
       "public_repo",
       "created_at",
       "operation_type",
